Prefer cancelling one permission prompt at one time, not all
When devs from the outside cancel all prompts, they may not consider that other
prompts were shown under different conditions.
Thus, each one should be cancelled independently.

The only case, where all of them can be cancelled (when the permission is granted)
is handled automatically and thus we deprecate the old function. (and renamed it)